Upload Combined Raw Data
Data was combined from individual data sheets recorded after each sampling period. NAs and blanks were removed because it's unclear if they were inputted because there was no reading to take (due to area dry, frozen, COVID, etc.), if sample was taken but failed, if sample was meant to be a zero but mistakenly recorded as NA or blank, or other complications not noted. Any recorded zero values were kept, assuming these were true and meant to represent a zero value for the sample.